client_api: fixes to version numbering stuff
Including not calling 16 bit quantities "nibbles".

In theory int can be less than 32 bits (although we certainly don't
support that in mpv), but using long here is still cleaner.

+++ player/client.c
@@ -673,7 +673,7 @@ int mpv_request_log_messages(mpv_handle *ctx, const char *min_level)
     return 0;
 }
 
-unsigned int mpv_client_api_version(void)
+unsigned long mpv_client_api_version(void)
 {
     return MPV_CLIENT_API_VERSION;
 }
